+++ instrumentation/tests.bzl
@@ -1,0 +1,9 @@
+"""
+Contains all the end-to-end test classes under the instrumentation directory.
+"""
+
+# Note to developers: Please keep this dict sorted by key to make it easier to find dependencies.
+# This list should contain classname and library target name of each test suite.
+END_TO_END_TEST_SUITES = {
+    "ExplorationPlayerTest": "//instrumentation/src/javatest/org/oppia/android/instrumentation:exploration_player_test_lib",
+}
